refactor: 지역 마커 렌더링 훅 변경
[#914]
