+++ frontend/src/components/google-maps/marker/components/SmallMediumDeltaAreaMarkerContainer/components/marker/CarffeineMarkerRenderer.tsx
@@ -8,7 +8,6 @@ interface Props {
   station: StationMarker;
 }
 
-// TODO: 여기다가 옵셔널로 infoWindow 열리고 닫히는 속성 넣어서 한번 컨트롤 해보기
 const CarffeineMarkerRenderer = ({ station }: Props) => {
   const { renderCarffeineMarker } = useMarker();
